Fix wrap-guide initialization bug
It seems to me that the author of this code assumed that `atom.workspace.observeTextEditor` calls the callback when a text editor gets created.
Looking through the code of the the portotype of `Workspace` I found that `Workspace.prototype.observeTextEditor` doesn't notify the creation of a text editor, it notifies the inclusion of if in the workspace instead.
A text editor can be removed from a workspace and afterwards included again, without destroying it. An item, for example, can be moved from one pane to another using the method `Pane.prototype.moveItemToPane`, which does the things I told previously, this action triggers the callbacks of `atom.workspace.observeTextEditor` and leads to an additional unnecessary initialization of another wrap-guide. A repetition of this action will lead to the creation multiple wrap-guide elements in the same text editor. For me this faulty behaviour is really annoying because the color of the wrap-guides on my editor is transparent, and when this bug occurs the line gets more opaque (because there are multiple wrap guides overlapping).
My solution for this issue is using a private weak set that keeps track of editors where a wrap guide has already been initialized; this set holds references to the editors weakly, therefore there shouldn't be any memory leakage.