Count the number of editors (not panes) when testing wrap guide count
atom/atom#13977 adds multiple pane locations and updates the
atom.workspace.getPanes() API to return the panes from all of them.
This fixes the resultant test failures by making assertions about the
number of text editors instead of the number of panes (which I think is
more true to the intent of the test).